Removing transform data
